refactored and change the use of the string

+++ chaos/common/action/DomainActions.cpp
@@ -80,7 +80,7 @@ string& DomainActions::getDomainName() {
 }
 
 uint32_t DomainActions::registeredActions() {
-	return actionDescriptionsMap.size();
+	return (uint32_t)actionDescriptionsMap.size();
 }
 
 /*
